#2: working on adding some more detailed documentation
