+++ docs-md/markugen.json
@@ -1,4 +1,6 @@
 [
   { "name" : "index", "title": "Home" },
-  { "name" : "Components", "toc": 4 }
+  { "name" : "Options" },
+  { "name" : "Components", "toc": 4 },
+  { "name" : "Styles", "toc": 4 }
 ]
